feat: add Sentry initialization condition and implement rate limiting for admin routes
